LPS-112093 Save the content associated to the collection in a new property in the fragmentEntryLink state.
Add two hooks to encapsulate the collection logic. one that call the render service and add the classPK and className of the item if it is in a collection. The other hook will get the content of the fragment, accessing the correct index based on the context information

If the element is not within a collection the default behaviour is preserved